Structural changes to the software FSM model
The framework for the software FSM has been simplified to use a single
struct that is passed between states (or methods). The RunFSM method
simply continues to call the Transit method which determines which state
handler to call based on the transition struct. Each state handler
returns a tansition struct indicating what state to go to next and what
events were registered.