RWStore uses vector instead of queue; set search
Changed the back end of the RWStore class to use a vector rather than a
queue. The operations exposed to the user of the class will still mimic
a queue, but internal operation require the functionality provided by the
vector class (primarily element accessors).

Added prototypes for methods that search the set vectors for addresses.
Mostly unimplemented.