Ready state returns ack state; commenting
The ready state just mimics the hardware waiting on a register event.
For simplicity, this software version is not event driven, so the state
machine must be executed explicitly. Therefor the ready state simply
transitions to the ack state at this point, it may also do some member
variable sanitization in the future.