Bit operation fix; States implemented; constructor
Added another bit to the operation bits in the control register giving
3-bits per control and the remaining 5 bits for the transaction id.

The states had some features that are required added to them and the
RunFSM method returns a bool to easily determine if it was an accept or
abort. Changed setCtrlAbort() to setCtrlOperation to allow easy seting
of the upper 3 bits (operation).

With the 3 bits of operation we can now indicate more precisely what
happened, therefore, accepted was added as a possible output from the
machine to indicate that the operation was allowed.